Backend/requirements.txt
# Core
fastapi==0.135.3
uvicorn[standard]==0.44.0
python-multipart==0.0.24
pydantic==2.12.5

# LLM & RAG
langchain==1.2.15
langchain-community==0.4.1
langchain-text-splitters==1.1.1
groq==1.1.2

# Vector Store
faiss-cpu==1.13.2

# Document Processing
PyMuPDF==1.27.2.2
Pillow==12.2.0
pytesseract==0.3.13

# ML / Classification
transformers==5.5.0
torch==2.11.0
sentencepiece==0.2.1
sentence-transformers==5.3.0

# Utilities
python-dateutil==2.9.0.post0
aiofiles==25.1.0
python-dotenv==1.2.2
supabase==2.11.0
